Moving get_storage_backend
